+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 ## 0.1.2
 
+### Futures
+
+- Add `Scheduler::run_sync`. To use it, add the `sync` flag to features.
+
+### Refactor
+
+- Rename `TaskCreator` to `ReactiveTask`, and rename `TaskCreator::task` to `ReactiveTask::will`
+- Delete `wait::while_` 
+
 ### Bug Fix
 
-- Delete `wait::while_`
 - Fixed wrong exit condition for `wait::until` (changed to wait until it returns true)